refactor: runner use DI for electron + gist setup
